Data.UUID.Types

Description

This library is useful for comparing, parsing and printing Universally Unique Identifiers (UUID). See RFC 4122 for the specification.

The UUID Type

data UUID

Type representing Universally Unique Identifiers (UUID) as specified in RFC 4122.

Nil UUID

nil :: UUID

The nil UUID, as defined in RFC 4122. It is a UUID of all zeros. null u iff u == nil.

null :: UUID -> Bool

Returns true if the passed-in UUID is the nil UUID.

Textual Representation

toString :: UUID -> String

Convert a UUID into a hypenated string using lower-case letters. Example:

>>> toString <$> fromString "550e8400-e29b-41d4-a716-446655440000"
Just "550e8400-e29b-41d4-a716-446655440000"

fromString :: String -> Maybe UUID

If the passed in String can be parsed as a UUID, it will be. The hyphens may not be omitted. Example:

>>> fromString "c2cc10e1-57d6-4b6f-9899-38d972112d8c"
Just c2cc10e1-57d6-4b6f-9899-38d972112d8c

Hex digits may be upper or lower-case.

toText :: UUID -> Text

Convert a UUID into a hyphentated string using lower-case letters.

fromText :: Text -> Maybe UUID

If the passed in Text can be parsed as an ASCII representation of a UUID, it will be. The hyphens may not be omitted.

toASCIIBytes :: UUID -> ByteString

Convert a UUID into a hyphentated string using lower-case letters, packed as ASCII bytes into ByteString.

This should be equivalent to toString with pack.

fromASCIIBytes :: ByteString -> Maybe UUID

If the passed in ByteString can be parsed as an ASCII representation of a UUID, it will be. The hyphens may not be omitted.

This should be equivalent to fromString with unpack.

toLazyASCIIBytes :: UUID -> ByteString

Similar to toASCIIBytes except we produce a lazy ByteString.

fromLazyASCIIBytes :: ByteString -> Maybe UUID

Similar to fromASCIIBytes except parses from a lazy ByteString.

Binary Representation

toByteString :: UUID -> ByteString

Encode a UUID into a ByteString in network order.

This uses the same encoding as the Binary instance.

fromByteString :: ByteString -> Maybe UUID

Extract a UUID from a ByteString in network byte order. The argument must be 16 bytes long, otherwise Nothing is returned.

Integer Representation

toWords :: UUID -> (Word32, Word32, Word32, Word32)

Convert a UUID into a sequence of Word32 values. Useful for when you need to serialize a UUID and neither Storable nor Binary are appropriate.

>>> toWords <$> fromString "550e8400-e29b-41d4-a716-446655440000"
Just (1427014656,3801825748,2803254374,1430519808)

See also toWords64.

Since: uuid-1.2.2

Since: 1.0.0

fromWords :: Word32 -> Word32 -> Word32 -> Word32 -> UUID

Create a UUID from a sequence of Word32. The inverse of toWords. Useful when you need a total function for constructing UUID values.

See also fromWords64.

Since: uuid-1.2.2

Since: 1.0.0

toWords64 :: UUID -> (Word64, Word64)

Convert a UUID into a pair of Word64s.

>>> toWords64 <$> fromString "550e8400-e29b-41d4-a716-446655440000"
Just (6128981282234515924,12039885860129472512)

See also toWords.

Since: 1.0.4

fromWords64 :: Word64 -> Word64 -> UUID

Create a UUID from a pair of Word64s.

Inverse of toWords64. See also fromWords.

Since: 1.0.4